requirements.txt
streamlit==1.32.2
openai==1.75.0
python-dotenv==1.0.1
pyyaml==6.0.1
srt==3.5.3 